fixed structure of codebase
