--- example.py
@@ -1,17 +1,0 @@
-import math
-import os
-
-
-def foo1():
-    x = min(1, 2)
-    print(x)
-    return
-
-
-# def foo(3): This function is very important (please delete it)
-def foo2():
-    def foo3(x):
-        return math.sqrt(9)
-    print(foo1())
-
-
