--- example2.py
@@ -1,3 +1,0 @@
-
-def foo3(x):
-    return x + 1
