+++ tests/example2.py
@@ -1,0 +1,3 @@
+
+def foo3(x):
+    return x + 1
